Add support for mention_slug parameter
Currently, the protocol assumes that the GitHub login supplied will
always be the appropriate username to use to refer to the user in chat
messages. This isn't always a safe assumption.

The motivating factor for this change is Slack's recent change to
usernames: https://api.slack.com/changelog/2017-09-the-one-about-usernames
However, it's also possible to imagine other systems in which a user's
GitHub username is distinct from their chat username; providing this
field should support a wider variety of possible usecases.