Add support for "attachment" feature
This adds support for a new option, which allows servers to hint
to clients that they should format the response as a separate
attachment instead of returning it inline as a message.
Because this is a hint, and because not all clients will support
an attachment mode, clients are allowed to ignore this field for
any reason.

For example, Slack supports "snippets", which are arbitrary text
documents sent as attachments. These don't use Slack's normal message
formatting and have a larger filesize limit.