Reduce the dependencies of the chatops controller
When looking at options to mitigate
https://github.com/github/github/issues/174690, my initial approach
there failed because we can't drop some Rails components because we use
gems that express all of Rails as a dependency instead of the parts they
need.

The chatops controller is one of those. It looks like though it only
needs ActiveSupport & ActionPack so I reduced the dependencies here to
just those parts.